Add system namespaces list
