+++ components/console-backend-service/internal/gqlschema/config.yml
@@ -121,3 +121,5 @@ models:
     fields:
       applications:
         resolver: true
+  Pod:
+    model: "github.com/kyma-project/kyma/components/console-backend-service/internal/gqlschema.Pod"
