+++ components/console-backend-service/internal/gqlschema/models_gen.go
@@ -382,18 +382,6 @@ type NavigationNode struct {
 	RequiredPermissions []RequiredPermission `json:"requiredPermissions"`
 }
 
-type Pod struct {
-	Name              string           `json:"name"`
-	NodeName          string           `json:"nodeName"`
-	Namespace         string           `json:"namespace"`
-	RestartCount      int              `json:"restartCount"`
-	CreationTimestamp time.Time        `json:"creationTimestamp"`
-	Labels            Labels           `json:"labels"`
-	Status            PodStatusType    `json:"status"`
-	ContainerStates   []ContainerState `json:"containerStates"`
-	JSON              JSON             `json:"json"`
-}
-
 type PodEvent struct {
 	Type SubscriptionEventType `json:"type"`
 	Pod  Pod                   `json:"pod"`
